xrpl-ipfs/requirements.txt
xrpl-py==4.3.1
base58==2.1.1
pynacl==1.6.1
python-dotenv==1.2.1